Un pochettino di refactoring
