+++ Assets/Scripts/Utils/Utils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 68eae9af33b87b845a4030e95ae84b56
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
